HORIZON-11: Incorporate EEG chart view
Revised the chart loader to be part of the charting section of the code folder.

Additionally, the chart loader class accepts a callback function rather than a post messages. Thus, it's not bound to a web-worker context. Making it simpler to validate in unit tests.

Whereas the worker for the chart loader is it's own code under the workers section.
