+++ src/charting/chart_loader/index.ts
@@ -1,0 +1,2 @@
+export * from './chart_loader';
+export * from './types';
